Correct issue for deleted files

+++ AOTMap/MACCSAOTMap.py
@@ -94,4 +94,5 @@ class MACCSAOTMap:
 
             logging.info('AOT image for resolution %sm : %s' %(s_resol,s_AOTPath))
 
-            os.remove(s_AOTPath + '.aux.xml')
+            if os.path.exists(s_AOTPath + '.aux.xml'):
+                os.remove(s_AOTPath + '.aux.xml')
