+++ waterVapor/MuscateWaterVapor.py
@@ -88,4 +88,5 @@ class MuscateWaterVapor:
 
             logging.info('Water vapor image for resolution %sm : %s' %(s_resol,s_WVPPath))
 
-            os.remove(s_WVPPath + '.aux.xml')
+            if os.path.exists(s_WVPPath + '.aux.xml'):
+                os.remove(s_WVPPath + '.aux.xml')
